Moving APIs from HostInfoLinux to HostInfoPosix
